Fix: End highlighting with only one word in text
If there is only one word in the text and that
word is a keyword, the highlighting only stops
on the next iteration (Stop character is added in else part,
which is never activated if there is only a or multiple keywords).

Adds a validation if the text is only a word long and if that
word is a keyword, if so, adds the end character to the text.