Add some testes
Add tests for some basic functions, the test situation is still not very
good, but its not the primary focus of this project right now. So to
have a passing nox result, I reduce the threshold for test coverage
to 25 percent. Maybe I will have time in the future to add more tests.