Fix highligth overflow #1
If the last word in a text is a keyword, the else
branch was not visited again after it. So the
Style.RESET_ALL flag was never added to the styled
text and everything that was printed afterwards
was also highlighted. The highlighting only stopped,
when a the else branch was reached for another text.
The code now checks if the highlighted word is the last
one for this text and if so, adds the Style.RESET_ALL
flag after the word.

Fixes: #1